Move _get_fs_and_path to progress tracker since it's the only use
